refactor (spelling): Исправлена опечатка в наименовании файлов
