+++ src/main.js
@@ -1,7 +1,7 @@
 import {createSiteMenuTemplate} from "./components/site-menu.js";
 import {generateFilters} from "./mock/filter.js";
 import {createFilterTemplate} from "./components/filter.js";
-import {generateTasks} from "./mock/taks.js";
+import {generateTasks} from "./mock/task.js";
 import {createTaskTemplate} from "./components/task.js";
 import {createTaskEditTemplate} from "./components/task-editor.js";
 import {createBoardTemplate} from "./components/board.js";
